fix: better error reporting
