+++ bloomstack_core/hook_events/user.py
@@ -39,7 +39,7 @@ def execute_bloomtrace_integration_request():
 			integration_request.error = ""
 			integration_request.status = "Completed"
 		except Exception as e:
-			integration_request.error = cstr(e)
+			integration_request.error = cstr(frappe.get_traceback())
 			integration_request.status = "Failed"
 
 		integration_request.save(ignore_permissions=True)
